refactor(ui): move all helper functions into multiple files
