+++ ui/src/helpers/ellipsisUnnecessaryParams.ts
@@ -1,0 +1,5 @@
+export default function ellipsisUnnecessaryParams(jsonString: string): string {
+  const cleanedJsonString = jsonString.replace(/"params":[^}]*}/g, '"params": { ... }')
+
+  return cleanedJsonString
+}
