+++ ui/src/helpers/parsedExpectedValue.ts
@@ -1,0 +1,11 @@
+import extractFieldValue from '@/helpers/extractFieldValue'
+
+export default function parsedExtractFieldValue(jsonString: string, fieldName: string): string | undefined {
+  const fieldValue = extractFieldValue(jsonString, fieldName)
+
+  if (fieldValue === undefined) {
+    return
+  }
+
+  return JSON.parse(fieldValue)
+}
