+++ ui/src/helpers/splitList.ts
@@ -1,0 +1,7 @@
+export default function splitList(list: string): string[] {
+  try {
+    return list.split(',').map((item) => item.trim()).sort()
+  } catch (e) {
+    return [list]
+  }
+}
